+++ scalikejdbc-mapper-generator-core/src/main/scala/scalikejdbc/mapper/GeneratorTemplate.scala
@@ -1,0 +1,8 @@
+package scalikejdbc.mapper
+
+case class GeneratorTemplate(name: String)
+
+object GeneratorTemplate {
+  val interpolation = GeneratorTemplate("interpolation")
+  val queryDsl = GeneratorTemplate("queryDsl")
+}
