+++ scalikejdbc-mapper-generator-core/src/main/scala/scalikejdbc/mapper/GeneratorTestTemplate.scala
@@ -1,0 +1,9 @@
+package scalikejdbc.mapper
+
+object GeneratorTestTemplate {
+  val ScalaTestFlatSpec = GeneratorTestTemplate("ScalaTestFlatSpec")
+  val specs2unit = GeneratorTestTemplate("specs2unit")
+  val specs2acceptance = GeneratorTestTemplate("specs2acceptance")
+}
+
+case class GeneratorTestTemplate(name: String)
